requirements.txt
pandas
gspread
oauth2client
python-dotenv
openpyxl
facebook_business